refactor: Replace RedirectAllHttpHandler usage with RedirectingHttpHandler
